refactor: move compatibility checks
